Fixed mistake where index and not ID was grabbed
When looking for a dialogue that can be shown it was correctly looking through the whole table and grabbed the corresponding ID. However, I then assumed that ID === index. Which could very easily not be the case. It worked out fine for a while until I caught that mistake now. Now it immediately grabs a copy of that table index and works with that instead