Sound folder changes + fixed dialogue sound bug
Dialogue mp3s are now in "dataAddon/audio". There are no more subfolders for the NPCs.

This made me look at the PlayAudio code again and I found and fixed the bug that stopped audio files from being skipped when clicking the next dialogue choice for a dialogue that has no audio.