Added 3D NPC portrait to dialogue
The 3D dialogue portrait uses the talk animation for roughly as long as it takes to read the text aloud. So the talk emote looping is not bound to the sound but to the length of the text.

Also fixed a bunch of stuff that was broken in the new wow addon version after adding new features